Remove unnecessary indirection in s3boto3 backend
The indirection was necessary for the older s3boto backend as it shared
code with the (now deprecated) GS backend. The s3boto3 does not share
code in this way. When the s3boto3 was developed, the indirection was
carried over, but it is no longer necessary. To simplify the code,
remove it. Other backends do not encode this style of indirection.